backend/requirements.txt
fastapi
uvicorn
httpx
anthropic
reportlab
python-dotenv
pydantic
